Upgrades in setup_contest.sh and setup_problem.sh. Now we automatically open all the files in the right orders. Also the number of problems to set up is an argument now, defaults to 5.

+++ templates/setup_contest.sh
@@ -2,10 +2,19 @@
 
 template_dir="$(dirname $(realpath "$0"))"
 
-if [ $# -ne 1 ]
+# Default number of problems to set up.
+NUM_PROBLEMS=5
+
+# Check that the number of arguments is either 1 or 2.
+if [ $# -eq 0 -o $# -gt 2 ]
 then
-    echo "Need exactly one argument (contest folder name)."
+    echo "Need 1-2 arguments (contest folder name, num problems defaults to $NUM_PROBLEMS)."
     exit
+fi
+
+if [ $# -eq 2 ]
+then
+    NUM_PROBLEMS=$2
 fi
 
 mkdir -p "$1"
@@ -14,11 +23,51 @@ cp "$template_dir/compare.sh" "$1/"
 # cp "$template_dir/compare.py" "$1/"
 cp "$template_dir/interact.sh" "$1/"
 cp "$template_dir/interact.py" "$1/"
-for PROBLEM in A B C D E F G H I J K L M N O P Q R S T
+
+# Keep a counter of how many problems we've set up,
+# so we can break once we've set up NUM_PROBLEMS.
+problem_counter=0
+
+# Also, maintain a list of all the files we've created so we can open them all at once at the end.
+files_created=""
+
+for PROBLEM in A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
 do
-    cp "$template_dir/template.py"  "$1/$PROBLEM.py"
-    cp "$template_dir/template.cpp" "$1/$PROBLEM.cpp"
-    touch "$1/$PROBLEM.in"
-    touch "$1/$PROBLEM.thinking"
+    # Break if we've set up enough problems.
+    if [ $problem_counter -eq $NUM_PROBLEMS ]
+    then
+        break
+    fi
+    problem_counter=$((problem_counter+1))
+
+    # For each file we're about to create, first check if it already exists.
+    # If so, don't overwrite it.
+
+    if [ -f "$1/$PROBLEM.thinking" ]
+    then
+        echo "File $1/$PROBLEM.thinking already exists. Skipping."
+    else
+        touch "$1/$PROBLEM.thinking"
+    fi
+    files_created="$files_created $1/$PROBLEM.thinking"
+
+    if [ -f "$1/$PROBLEM.cpp" ]
+    then
+        echo "File $1/$PROBLEM.cpp already exists. Skipping."
+    else
+        cp "$template_dir/template.cpp"  "$1/$PROBLEM.cpp"
+    fi
+    files_created="$files_created $1/$PROBLEM.cpp"
+
+    if [ -f "$1/$PROBLEM.in" ]
+    then
+        echo "File $1/$PROBLEM.in already exists. Skipping."
+    else
+        touch "$1/$PROBLEM.in"
+    fi
+    files_created="$files_created $1/$PROBLEM.in"
 done
-cd "$1"
+
+# Open the folder $1 in VS Code, and open all the files we created.
+code -n "$1"
+code -r $files_created
